BGDIINF_SB-1261: Fixed dockerrun target
The dockerrun make target was done to run a local docker image on port
8080. The port 8080 is already used by the BIT proxy on the developper
machines and there were not target for building a local image.

Now there is a target to build local image that can be started by
dockerrun.

The default exported port by the local docker image is 5000 and can be
changed by an environment variable. This variable needs to be set prior
to deployment in order to use another port, but as deployment will be
automated it is not an issue to have to  set an environment variable
before it.