Fixed route logging
The route logging middleware did not add the 'request' info object in
json message as configured with the Flask Filter. This was because in
the middleware there is no Flask request context yet.

Therefore moved the log middleware into a before_request method.

Also removed the dev and prod logging configuration that should be part
of the deployment configuration (k8s)